setup.sh修正

+++ setup.sh
@@ -4,6 +4,7 @@ cd ~/.dotfiles/files/
 
 for file in `ls -a`
 do
-    ln -s $HOME/.dotfiles/files/$file $HOME/$file
+    if [ ! -e $HOME/$file ]; then
+        ln -s $HOME/.dotfiles/files/$file $HOME/$file
+    fi
 done
-mv ~/dotfiles ~/.dotfiles
